pyproject.toml
[project]
name = "cal26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.2",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "pillow>=11.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qrcode>=8.0",
    "starlette>=0.50.0",
    "uvicorn>=0.38.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"
